fix: harden terminal restore review gaps
